moving ProjectEntry stuff to another file
